+++ platforms/react/src/components/ui/TextArea/index.ts
@@ -1,7 +1,0 @@
-import { withTextShallot } from '@/utils'
-
-export const TextArea = withTextShallot('textarea', {
-  display: 'flex',
-  width: '100%',
-  border: '1px solid transparent',
-})