+++ components/Modal/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @shallot-ui/component-modal
 
+## 2.1.0
+
+### Minor Changes
+
+- Adds textarea
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/core-mixins@2.1.0
+  - @shallot-ui/core-theme@2.1.0
+  - @shallot-ui/core-utils@2.1.0
+
 ## 2.0.1
 
 ### Patch Changes