+++ core/mixins/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @shallot-ui/core-mixins
 
+## 2.1.0
+
+### Minor Changes
+
+- Adds textarea
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/core-theme@2.1.0
+
 ## 2.0.1
 
 ### Patch Changes