+++ packages/web/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @shallot-ui/web
 
+## 4.1.0
+
+### Minor Changes
+
+- Adds textarea
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/component-checkbox@3.1.0
+  - @shallot-ui/component-button@3.1.0
+  - @shallot-ui/component-switch@3.1.0
+  - @shallot-ui/component-input@3.1.0
+  - @shallot-ui/component-modal@2.1.0
+  - @shallot-ui/platform-react@2.1.0
+  - @shallot-ui/component-tag@3.1.0
+  - @shallot-ui/core-mixins@2.1.0
+  - @shallot-ui/core-props@2.1.0
+  - @shallot-ui/core-theme@2.1.0
+  - @shallot-ui/core-utils@2.1.0
+
 ## 4.0.4
 
 ### Patch Changes