+++ platforms/nextjs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @shallot-ui/platform-nextjs
 
+## 2.1.0
+
+### Minor Changes
+
+- Adds textarea
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/platform-react@2.1.0
+  - @shallot-ui/core-props@2.1.0
+  - @shallot-ui/core-theme@2.1.0
+
 ## 2.0.4
 
 ### Patch Changes